+++ tests/graph/test_connector.py
@@ -30,7 +30,7 @@ def mocked_query_class(monkeypatch: MonkeyPatch) -> None:
     monkeypatch.setattr(Query, "render", lambda s: call(s.name, **s.kwargs))
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_check_connectivity_and_authentication(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [{"currentStatus": "online"}]
     graph = GraphConnector.get()
@@ -39,7 +39,7 @@ def test_check_connectivity_and_authentication(mocked_graph: MockedGraph) -> Non
     assert mocked_graph.call_args_list[-1] == call(call("fetch_database_status"), {})
 
 
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_check_connectivity_and_authentication_error(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [{"currentStatus": "offline"}]
     graph = GraphConnector.get()
@@ -47,7 +47,7 @@ def test_check_connectivity_and_authentication_error(mocked_graph: MockedGraph) 
         graph._check_connectivity_and_authentication()
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_seed_constraints(mocked_graph: MockedGraph) -> None:
     graph = GraphConnector.get()
     graph._seed_constraints()
@@ -79,7 +79,7 @@ def test_mocked_graph_seed_constraints(mocked_graph: MockedGraph) -> None:
     )
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_seed_indices(
     mocked_graph: MockedGraph, monkeypatch: MonkeyPatch
 ) -> None:
@@ -141,7 +141,7 @@ def test_mocked_graph_seed_indices(
     )
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_seed_data(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [
         {
@@ -213,7 +213,7 @@ def test_mocked_graph_seed_data(mocked_graph: MockedGraph) -> None:
     )
 
 
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_mocked_graph_commit_raises_error(mocked_graph: MockedGraph) -> None:
     mocked_graph.run.side_effect = Exception("query failed")
     graph = GraphConnector.get()
@@ -221,7 +221,7 @@ def test_mocked_graph_commit_raises_error(mocked_graph: MockedGraph) -> None:
         graph._check_connectivity_and_authentication()
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_fetch_extracted_items(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [
         {
@@ -506,7 +506,7 @@ def test_fetch_extracted_items(
     assert result.one() == expected
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_fetch_rule_items(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [
         {
@@ -663,7 +663,7 @@ def test_fetch_rule_items_empty() -> None:
     assert result.one() == {"items": [], "total": 0}
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_fetch_merged_items(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [
         {
@@ -759,7 +759,7 @@ def test_mocked_graph_fetch_merged_items(mocked_graph: MockedGraph) -> None:
     }
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis", "mocked_graph")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey", "mocked_graph")
 def test_mocked_graph_fetch_merged_items_invalid_field_name() -> None:
     graph = GraphConnector.get()
     with pytest.raises(ValueError, match="Invalid field name"):
@@ -1248,7 +1248,7 @@ def test_fetch_merged_items(  # noqa: PLR0913
     assert result.one() == expected
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_fetch_identities(mocked_graph: MockedGraph) -> None:
     graph = GraphConnector.get()
     graph.fetch_identities(stable_target_id=Identifier.generate(99))
@@ -1306,7 +1306,7 @@ def test_mocked_graph_fetch_identities(mocked_graph: MockedGraph) -> None:
     )
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_exists_item(
     mocked_graph: MockedGraph,
     monkeypatch: MonkeyPatch,
@@ -1357,7 +1357,7 @@ def test_graph_exists_item(
     assert graph.exists_item(stable_target_id, entity_types) == exists
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_run_ingest_in_transaction_ruleset(
     mocked_graph: MockedGraph,
     organizational_unit_rule_set_response: OrganizationalUnitRuleSetResponse,
@@ -1411,7 +1411,7 @@ def test_graph_merge_rule_edges_fails_inconsistent(
         deque(graph.ingest_items([consistent_org, inconsistent_unit]))
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_ingests_rule_set(
     mocked_graph: MockedGraph,
     organizational_unit_rule_set_response: OrganizationalUnitRuleSetResponse,
@@ -1429,7 +1429,7 @@ def test_mocked_graph_ingests_rule_set(
     )
 
 
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_mocked_graph_ingests_extracted_models(
     mocked_graph: MockedGraph,
     dummy_data: dict[str, AnyExtractedModel],
@@ -1830,7 +1830,7 @@ def test_connector_flush_fails(monkeypatch: MonkeyPatch) -> None:
         graph.flush()
 
 
-@pytest.mark.usefixtures("mocked_query_class", "mocked_redis")
+@pytest.mark.usefixtures("mocked_query_class", "mocked_valkey")
 def test_mocked_graph_delete_item(mocked_graph: MockedGraph) -> None:
     mocked_graph.return_value = [
         {