+++ tests/identity/test_provider.py
@@ -52,7 +52,7 @@ from tests.conftest import MockedGraph
     ],
     ids=["new item", "existing item"],
 )
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_assign_identity_mocked(
     mocked_graph: MockedGraph,
     mocked_return: list[dict[str, str]],
@@ -69,7 +69,7 @@ def test_assign_identity_mocked(
     assert identity.model_dump() == expected
 
 
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_assign_identity_inconsistency_mocked(
     mocked_graph: MockedGraph,
 ) -> None:
@@ -213,7 +213,7 @@ def test_assign_identity(
     ],
     ids=["nothing found", "one item", "two items"],
 )
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_fetch_identities_mocked(  # noqa: PLR0913
     mocked_graph: MockedGraph,
     mocked_return: list[dict[str, str]],