+++ tests/merged/test_main.py
@@ -18,7 +18,7 @@ from mex.common.types import Validation
 from tests.conftest import MockedGraph
 
 
-@pytest.mark.usefixtures("mocked_redis")
+@pytest.mark.usefixtures("mocked_valkey")
 def test_search_merged_items_mocked(
     client_with_api_key_read_permission: TestClient,
     mocked_graph: MockedGraph,