+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changes
+
 - bumped cookiecutter template to https://github.com/robert-koch-institut/mex-template/commit/3c389d
+- **BREAKING** replace redis cache with valkey
 
 ### Deprecated